♻️|Add messagesPreviewFragment/ refactor di modules
