+++ app/src/main/java/com/alibasoglu/ciftlikpazarimandroid/FavoritesFragment.kt
@@ -3,7 +3,9 @@ package com.alibasoglu.ciftlikpazarimandroid
 import com.alibasoglu.ciftlikpazarimandroid.core.fragment.BaseFragment
 import com.alibasoglu.ciftlikpazarimandroid.core.fragment.FragmentConfiguration
 import com.alibasoglu.ciftlikpazarimandroid.core.fragment.ToolbarConfiguration
+import dagger.hilt.android.AndroidEntryPoint
 
+@AndroidEntryPoint
 class FavoritesFragment : BaseFragment(R.layout.fragment_favorites) {
 
     private val toolbarConfiguration = ToolbarConfiguration(R.string.favorites)
