+++ app/src/main/java/com/alibasoglu/ciftlikpazarimandroid/adverts/di/AdvertModule.kt
@@ -1,17 +1,14 @@
 package com.alibasoglu.ciftlikpazarimandroid.adverts.di
 
-import android.content.SharedPreferences
 import com.alibasoglu.ciftlikpazarimandroid.adverts.data.AdvertsApi
 import com.alibasoglu.ciftlikpazarimandroid.adverts.data.AdvertsRepositoryImpl
 import com.alibasoglu.ciftlikpazarimandroid.adverts.domain.AdvertsRepository
-import com.alibasoglu.ciftlikpazarimandroid.auth.data.AuthRepositoryImpl
 import com.alibasoglu.ciftlikpazarimandroid.di.AppModule
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import javax.inject.Singleton
-import okhttp3.Interceptor
 import okhttp3.OkHttpClient
 import retrofit2.Retrofit
 import retrofit2.converter.moshi.MoshiConverterFactory
@@ -20,27 +17,6 @@ import retrofit2.create
 @Module
 @InstallIn(SingletonComponent::class)
 object AdvertModule {
-
-    @Provides
-    @Singleton
-    fun provideOkHttpClient(sharedPreferences: SharedPreferences): OkHttpClient {
-        val authToken = sharedPreferences.getString(AuthRepositoryImpl.AUTH_TOKEN_PREFERENCE_NAME, "") ?: "no Token"
-
-        val headerInterceptor = Interceptor { chain ->
-            chain.run {
-                proceed(
-                    request()
-                        .newBuilder()
-                        .addHeader("x-auth-token", authToken)
-                        .build()
-                )
-            }
-        }
-
-        return OkHttpClient.Builder()
-            .addInterceptor(headerInterceptor)
-            .build()
-    }
 
     @Provides
     @Singleton
