+++ app/src/main/java/com/alibasoglu/ciftlikpazarimandroid/di/AppModule.kt
@@ -12,6 +12,8 @@ import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import javax.inject.Singleton
+import okhttp3.Interceptor
+import okhttp3.OkHttpClient
 import retrofit2.Retrofit
 import retrofit2.converter.moshi.MoshiConverterFactory
 import retrofit2.create
@@ -20,6 +22,27 @@ import retrofit2.create
 @Module
 @InstallIn(SingletonComponent::class)
 object AppModule {
+
+    @Provides
+    @Singleton
+    fun provideOkHttpClient(sharedPreferences: SharedPreferences): OkHttpClient {
+        val authToken = sharedPreferences.getString(AuthRepositoryImpl.AUTH_TOKEN_PREFERENCE_NAME, "") ?: "no Token"
+
+        val headerInterceptor = Interceptor { chain ->
+            chain.run {
+                proceed(
+                    request()
+                        .newBuilder()
+                        .addHeader("x-auth-token", authToken)
+                        .build()
+                )
+            }
+        }
+
+        return OkHttpClient.Builder()
+            .addInterceptor(headerInterceptor)
+            .build()
+    }
 
     @Provides
     @Singleton
